requirements.txt
torch
torchvision
pandas
scikit-learn
Pillow